refactor: moved BidaDepartment type to ./config/types
