+++ src/containers/NodeProjectDependenciesTable.tsx
@@ -6,7 +6,7 @@ import Tag from '../components/Tag'
 import Table, { Column } from '../components/Table'
 import anchorRowRenderer from '../utils/anchorRowRenderer'
 import { isValidLicense } from '../data/Library/index'
-import { BidaDepartment } from '../data/__generated-types'
+import { BidaDepartment } from '../config/types'
 import * as routes from '../routes/routes'
 
 export interface Props {
