--- src/data/__generated-types.ts
@@ -1,20 +1,0 @@
-/* tslint:disable */
-// This file was automatically generated and should not be edited.
-
-//= =============================================================
-// START Enums and Input Objects
-//= =============================================================
-
-/**
- * Enumerator that indicates types of projects this application acknowledges
- */
-export enum BidaDepartment {
-  ANDROID = 'ANDROID',
-  BACKEND = 'BACKEND',
-  FRONTEND = 'FRONTEND',
-  IOS = 'IOS'
-}
-
-//= =============================================================
-// END Enums and Input Objects
-//= =============================================================
